+++ cmd/eulabeia-sensor/main.go
@@ -11,10 +11,11 @@ import (
 	"github.com/google/uuid"
 	"github.com/greenbone/eulabeia/connection/mqtt"
 	"github.com/greenbone/eulabeia/messages"
+	"github.com/greenbone/eulabeia/messages/cmds"
 )
 
 func main() {
-	topic := "greenbone.sensor"
+	// topic := "eulabeia/+/+/sensor"
 	server := flag.String("server", "localhost:1883", "A clientid for the connection")
 	clientid := flag.String("clientid", "", "A clientid for the connection")
 	sensorID := flag.String("sensorID", "bla", "A sensorID for the registration")
@@ -23,10 +24,12 @@ func main() {
 	log.Println("Starting sensor")
 	client, err := mqtt.New(*server, *clientid+uuid.NewString(), "", "",
 		&mqtt.LastWillMessage{
-			Topic: topic,
-			MSG: messages.Delete{
-				ID:      *sensorID,
-				Message: messages.NewMessage("delete.sensor", "", ""),
+			Topic: "eulabeia/sensor/cmd/director",
+			MSG: cmds.Delete{
+				Identifier: messages.Identifier{
+					Message: messages.NewMessage("delete.sensor", "", ""),
+					ID:      *sensorID,
+				},
 			}})
 	if err != nil {
 		log.Panicf("Failed to create MQTT: %s", err)
@@ -35,9 +38,11 @@ func main() {
 	if err != nil {
 		log.Panicf("Failed to connect: %s", err)
 	}
-	client.Publish(topic, messages.Modify{
-		Message: messages.NewMessage("modify.sensor", "", ""),
-		ID:      *sensorID,
+	client.Publish("eulabeia/sensor/cmd/director", cmds.Modify{
+		Identifier: messages.Identifier{
+			Message: messages.NewMessage("modify.sensor", "", ""),
+			ID:      *sensorID,
+		},
 		Values: map[string]interface{}{
 			"type": "undefined",
 		},