+++ director/scan/scan_test.go
@@ -5,24 +5,27 @@ import (
 
 	"github.com/greenbone/eulabeia/internal/test"
 	"github.com/greenbone/eulabeia/messages"
+	"github.com/greenbone/eulabeia/messages/cmds"
 	"github.com/greenbone/eulabeia/messages/handler"
 	"github.com/greenbone/eulabeia/storage"
 )
 
 func TestCreateScan(t *testing.T) {
-	h := handler.New(New("greenbone.sensor", storage.Noop{}))
+	h := handler.New(New(storage.Noop{}))
 	tests := []test.HandleTests{
 		{
-			Input: messages.Create{
+			Input: cmds.Create{
 				Message: messages.NewMessage("create.scan", "1", "1"),
 			},
 			Handler:         h,
 			ExpectedMessage: messages.NewMessage("created.scan", "1", "1"),
 		},
 		{
-			Input: messages.Start{
-				Message: messages.NewMessage("start.scan", "1", "1"),
-				ID:      "1234",
+			Input: cmds.Start{
+				Identifier: messages.Identifier{
+					Message: messages.NewMessage("start.scan", "1", "1"),
+					ID:      "1234",
+				},
 			},
 			Handler: h,
 			// although NoopStorage for target doesn't have sensor it should just
@@ -30,9 +33,11 @@ func TestCreateScan(t *testing.T) {
 			ExpectedMessage: messages.NewMessage("start.scan.", "1", "1"),
 		},
 		{
-			Input: messages.Modify{
-				Message: messages.NewMessage("modify.scan", "1", "2"),
-				ID:      "123",
+			Input: cmds.Modify{
+				Identifier: messages.Identifier{
+					Message: messages.NewMessage("modify.scan", "1", "2"),
+					ID:      "123",
+				},
 				Values: map[string]interface{}{
 					"finished":  []string{"1", "2"},
 					"target_id": "1",
@@ -42,9 +47,11 @@ func TestCreateScan(t *testing.T) {
 			Handler:         h,
 		},
 		{
-			Input: messages.Modify{
-				Message: messages.NewMessage("modify.scan", "1", "2"),
-				ID:      "123",
+			Input: cmds.Modify{
+				Identifier: messages.Identifier{
+					Message: messages.NewMessage("modify.scan", "1", "2"),
+					ID:      "123",
+				},
 				Values: map[string]interface{}{
 					"exclude":   []string{"1", "2"},
 					"target_id": 1,
@@ -54,9 +61,11 @@ func TestCreateScan(t *testing.T) {
 			ExpectedMessage: messages.NewMessage("failure.modify.scan", "1", "2"),
 		},
 		{
-			Input: messages.Get{
-				Message: messages.NewMessage("get.scan", "1", "2"),
-				ID:      "123",
+			Input: cmds.Get{
+				Identifier: messages.Identifier{
+					Message: messages.NewMessage("get.scan", "1", "2"),
+					ID:      "123",
+				},
 			},
 			ExpectedMessage: messages.NewMessage("got.scan", "1", "2"),
 			Handler:         h,