+++ director/sensor/sensor_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/greenbone/eulabeia/internal/test"
 	"github.com/greenbone/eulabeia/messages"
+	"github.com/greenbone/eulabeia/messages/cmds"
 	"github.com/greenbone/eulabeia/messages/handler"
 	"github.com/greenbone/eulabeia/storage"
 )
@@ -13,16 +14,18 @@ func TestSensor(t *testing.T) {
 	h := handler.New(New(storage.Noop{}))
 	tests := []test.HandleTests{
 		{
-			Input: messages.Create{
+			Input: cmds.Create{
 				Message: messages.NewMessage("create.sensor", "1", "1"),
 			},
 			Handler:         h,
 			ExpectedMessage: messages.NewMessage("created.sensor", "1", "1"),
 		},
 		{
-			Input: messages.Modify{
-				Message: messages.NewMessage("modify.sensor", "1", "2"),
-				ID:      "123",
+			Input: cmds.Modify{
+				Identifier: messages.Identifier{
+					Message: messages.NewMessage("modify.sensor", "1", "2"),
+					ID:      "123",
+				},
 				Values: map[string]interface{}{
 					"type": "openvas",
 				},
@@ -31,9 +34,11 @@ func TestSensor(t *testing.T) {
 			Handler:         h,
 		},
 		{
-			Input: messages.Get{
-				Message: messages.NewMessage("get.sensor", "1", "2"),
-				ID:      "123",
+			Input: cmds.Get{
+				Identifier: messages.Identifier{
+					Message: messages.NewMessage("get.sensor", "1", "2"),
+					ID:      "123",
+				},
 			},
 			ExpectedMessage: messages.NewMessage("got.sensor", "1", "2"),
 			Handler:         h,