+++ CHANGELOG.md
@@ -11,5 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add scan, sensor aggregate; extend sensor to register and deregister itself [5](https://github.com/greenbone/eulabeia/pull/5)
 - Add scanner logic, scheduler and openvas module [10](https://github.com/greenbone/eulabeia/pull/10)
 ### Changed
+- Split cmds and info messages into own module [8](https://github.com/greenbone/eulabeia/pull/8)
+- Normalized topic structure to `group/aggregate/event/destination`; setting topic based on return message rather than configuration [8](https://github.com/greenbone/eulabeia/pull/8)
 ### Fixed
 ### Removed